LPS-54009 Extract common logic
