+++ modules/apps/document-library/document-library-item-selector-web/src/com/liferay/document/library/item/selector/web/file/DLFileItemSelectorView.java
@@ -14,21 +14,9 @@
 
 package com.liferay.document.library.item.selector.web.file;
 
-import com.liferay.document.library.item.selector.web.DLItemSelectorView;
+import com.liferay.document.library.item.selector.web.BaseDLItemSelectorView;
 import com.liferay.item.selector.ItemSelectorView;
 import com.liferay.item.selector.criteria.file.criterion.FileItemSelectorCriterion;
-
-import java.io.IOException;
-
-import java.util.Locale;
-import java.util.ResourceBundle;
-
-import javax.portlet.PortletURL;
-
-import javax.servlet.RequestDispatcher;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -37,47 +25,11 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(service = ItemSelectorView.class)
 public class DLFileItemSelectorView
-	implements DLItemSelectorView<FileItemSelectorCriterion> {
-
-	public static final String DL_ITEM_SELECTOR_CRITERION =
-		DLFileItemSelectorView.class.getName() +
-			"#DL_FILE_ITEM_SELECTOR_CRITERION";
-
-	public static final String ITEM_SELECTED_EVENT_NAME =
-		DLFileItemSelectorView.class.getName() + "#ITEM_SELECTED_EVENT_NAME";
-
-	public static final String PORTLET_URL =
-		DLFileItemSelectorView.class.getName() + "#PORTLET_URL";
+	extends BaseDLItemSelectorView<FileItemSelectorCriterion> {
 
 	@Override
 	public Class<FileItemSelectorCriterion> getItemSelectorCriterionClass() {
 		return FileItemSelectorCriterion.class;
-	}
-
-	@Override
-	public String getTitle(Locale locale) {
-		ResourceBundle resourceBundle = ResourceBundle.getBundle(
-			"content/Language", locale);
-
-		return resourceBundle.getString("documents");
-	}
-
-	@Override
-	public void renderHTML(
-			ServletRequest request, ServletResponse response,
-			FileItemSelectorCriterion fileItemSelectorCriterion,
-			PortletURL portletURL, String itemSelectedEventName)
-		throws IOException, ServletException {
-
-		request.setAttribute(
-			DL_ITEM_SELECTOR_CRITERION, fileItemSelectorCriterion);
-		request.setAttribute(ITEM_SELECTED_EVENT_NAME, itemSelectedEventName);
-		request.setAttribute(PORTLET_URL, portletURL);
-
-		RequestDispatcher requestDispatcher = request.getRequestDispatcher(
-			"/o/document-library-item-selector-web/documents.jsp");
-
-		requestDispatcher.include(request, response);
 	}
 
 }
