+++ modules/apps/document-library/document-library-item-selector-web/src/com/liferay/document/library/item/selector/web/image/DLImageItemSelectorView.java
@@ -14,21 +14,9 @@
 
 package com.liferay.document.library.item.selector.web.image;
 
-import com.liferay.document.library.item.selector.web.DLItemSelectorView;
+import com.liferay.document.library.item.selector.web.BaseDLItemSelectorView;
 import com.liferay.item.selector.ItemSelectorView;
 import com.liferay.item.selector.criteria.image.criterion.ImageItemSelectorCriterion;
-
-import java.io.IOException;
-
-import java.util.Locale;
-import java.util.ResourceBundle;
-
-import javax.portlet.PortletURL;
-
-import javax.servlet.RequestDispatcher;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -37,47 +25,11 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(service = ItemSelectorView.class)
 public class DLImageItemSelectorView
-	implements DLItemSelectorView<ImageItemSelectorCriterion> {
-
-	public static final String DL_ITEM_SELECTOR_CRITERION =
-		DLImageItemSelectorView.class.getName() +
-			"#DL_IMAGE_ITEM_SELECTOR_CRITERION";
-
-	public static final String ITEM_SELECTED_EVENT_NAME =
-		DLImageItemSelectorView.class.getName() + "#ITEM_SELECTED_EVENT_NAME";
-
-	public static final String PORTLET_URL =
-		DLImageItemSelectorView.class.getName() + "#PORTLET_URL";
+	extends BaseDLItemSelectorView<ImageItemSelectorCriterion> {
 
 	@Override
 	public Class<ImageItemSelectorCriterion> getItemSelectorCriterionClass() {
 		return ImageItemSelectorCriterion.class;
-	}
-
-	@Override
-	public String getTitle(Locale locale) {
-		ResourceBundle resourceBundle = ResourceBundle.getBundle(
-			"content/Language", locale);
-
-		return resourceBundle.getString("documents");
-	}
-
-	@Override
-	public void renderHTML(
-			ServletRequest request, ServletResponse response,
-			ImageItemSelectorCriterion imageItemSelectorCriterion,
-			PortletURL portletURL, String itemSelectedEventName)
-		throws IOException, ServletException {
-
-		request.setAttribute(
-			DL_ITEM_SELECTOR_CRITERION, imageItemSelectorCriterion);
-		request.setAttribute(ITEM_SELECTED_EVENT_NAME, itemSelectedEventName);
-		request.setAttribute(PORTLET_URL, portletURL);
-
-		RequestDispatcher requestDispatcher = request.getRequestDispatcher(
-			"/o/document-library-item-selector-web/documents.jsp");
-
-		requestDispatcher.include(request, response);
 	}
 
 }
